refactor(cli): extract `glyph inspect` into its own module
`glyph_cmds.py` had grown to 1656 lines. The `inspect` subcommand was its
single largest, most self-contained feature — its core classifiers already
live in `pyrxd.glyph._inspect_core`, so the CLI side was only thin
error-translation wrappers, three human-mode renderers, the command, and the
`_run_fetch_inspect` network orchestration. Move all of it verbatim into a new
`pyrxd.cli.glyph_inspect` module (1214 + 473 lines; glyph_cmds.py drops 450).

The only behavioral-equivalent change: the command's decorator goes from
`@glyph_group.command(name="inspect")` to a bare `@click.command(name="inspect")`,
and `glyph_cmds.py` attaches it with `glyph_group.add_command(inspect_cmd)` — the
canonical Click pattern for splitting a group's subcommands across files. For a
plain `@click.group`, `Group.command()` injects no group state, so the two forms
produce an identical Command (verified: `glyph --help` / `glyph inspect --help`
byte-identical before/after).

Backward compat preserved: `inspect_cmd`, `_sanitize_display_string`, and
`_HUMAN_STRING_CAP` still import from `pyrxd.cli.glyph_cmds` (re-exported via the
`X as X` pattern), and `__all__` is unchanged. The Pyodide inspect façade
(`pyrxd.glyph.inspect` → `_inspect_core`) is untouched.

No logic changed: the 414 moved body lines diff to exactly the one decorator
line. ruff + format clean; full suite green (3734 passed); CLI verified
end-to-end in human/json/quiet modes.
